+++ plugin_ide.core/src/mmrnmhrm/core/DeeCorePreferencesConstants.java
@@ -10,11 +10,12 @@
  *******************************************************************************/
 package mmrnmhrm.core;
 
+import melnorme.lang.ide.core.utils.prefs.ProgramArgumentsPreference;
 import melnorme.lang.ide.core.utils.prefs.StringPreference;
 
 public interface DeeCorePreferencesConstants {
 	
 	static StringPreference PREF_DUB_PATH = new StringPreference("DUB_PATH", "dub");
-	static StringPreference DUB_BUILD_OPTIONS = new StringPreference("DUB_BUILD_OPTIONS", "");
+	static ProgramArgumentsPreference DUB_BUILD_OPTIONS = new ProgramArgumentsPreference("DUB_BUILD_OPTIONS", "");
 	
 }